re-organize webpack js files intto folders
