+++ webpack/js/fetch/fetch_data.js
@@ -6,7 +6,11 @@ const loginUrl = urlBase + 'api/authenticate';
 const logoutUrl = urlBase + 'auth/logout';
 const refreshUrl = urlBase + 'auth/refresh_token';
 const graphQLUrl = urlBase + 'graphql';
-const restUrlBase = urlBase + 'api/tree/';
+
+const apiEndpoints = {
+    tree: urlBase + 'api/tree/',
+    individuals: urlBase + 'api/individuals/'
+};
 
 const MEMBERSPACE_TOKEN = process.env.MEMBERSPACE_TOKEN || "MemberSpaceWidget.token"
 const ACCOUNT_PAGE = process.env.ACCOUNT_PAGE ||  "/account"
@@ -14,8 +18,10 @@ const ACCOUNT_PAGE = process.env.ACCOUNT_PAGE ||  "/account"
 const apiClient = new ApiClient();
 const errorHandler = new FetchErrorHandler(loginUrl, refreshUrl, logoutUrl, MEMBERSPACE_TOKEN, ACCOUNT_PAGE, apiClient);
 
-export async function getRestDataWrapper(url_end) {
-    var restUrl = restUrlBase + url_end;
+export async function getRestDataWrapper(endpoint, itemId) {
+    const restUrlBase = apiEndpoints[endpoint]
+    console.log(endpoint, itemId, restUrlBase)
+    const restUrl = restUrlBase + itemId;
 
     try {
         return await apiClient.getRestData(restUrl);
